backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Fullstack App - Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "PyJWT>=2.10.0",
]

[project.optional-dependencies]
dev = [
    "aiosqlite>=0.20.0",
    "httpx>=0.28.0",
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
